add parameter for script. finish all script example

+++ PythonScriptExample/bgm.py
@@ -19,6 +19,14 @@ level_dict = {
     "15": "5"
 }
 
+legal_character_dict = [
+    '1',
+    '2',
+    '3',
+    '4',
+    '5',
+]
+
 # game_path have slash
 # but current_folder don't have slash
 #
@@ -29,6 +37,17 @@ def install(game_path, current_folder):
 # Return true to report a successful config, otherwise return false
 def deploy(game_path, current_folder, parameter):
     try:
+        # restore old
+        deploy_cache = read_deploy(current_folder + "\\deploy.cfg")
+        if not deploy_cache == "":
+            target_file_1 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_1.wav"
+            target_file_2 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_2.wav"
+            target_file_3 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_3.wav"
+            remove_with_restore(target_file_1)
+            remove_with_restore(target_file_2)
+            remove_with_restore(target_file_3)
+
+        # deploy new
         theme = '1'
         parameter.split('_')
         if (parameter[0] == 'level'):
@@ -37,12 +56,12 @@ def deploy(game_path, current_folder, parameter):
                 return False
             theme = level_dict[parameter[1]]
         elif (parameter[0] == 'theme'):
-            test_theme = int(parameter[1])
-            if not (test_theme >= 1 and test_theme <= 5):
-                return False
             theme = parameter[1]
         else:
             return Flase
+
+        if not theme in legal_character_dict:
+            return False
         
         target_file_1 = game_path + "Sounds\\Music_Theme_" + theme + "_1.wav"
         target_file_2 = game_path + "Sounds\\Music_Theme_" + theme + "_2.wav"
@@ -50,25 +69,70 @@ def deploy(game_path, current_folder, parameter):
         copy_with_backups(target_file_1, current_folder + "\\1.wav")
         copy_with_backups(target_file_2, current_folder + "\\2.wav")
         copy_with_backups(target_file_3, current_folder + "\\3.wav")
+
+        record_deploy(current_folder + "\\deploy.cfg", theme)
     except:
         return False
     return True
 
 # Return true to report that package is intact, otherwise return false
 def check(game_path, current_folder):
+    deploy_cache = read_deploy(current_folder + "\\deploy.cfg")
+    if deploy_cache == "":
+        return True
+    target_file_1 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_1.wav"
+    target_file_2 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_2.wav"
+    target_file_3 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_3.wav"
+
+    if os.path.getsize(current_folder + "\\1.wav") != os.path.getsize(target_file_1):
+        return False
+    if os.path.getsize(current_folder + "\\2.wav") != os.path.getsize(target_file_2):
+        return False
+    if os.path.getsize(current_folder + "\\3.wav") != os.path.getsize(target_file_3):
+        return False
     return True
 
 # Return true to report that package is removed successfully, otherwise return false
-def remove(game_path):
+def remove(game_path, current_folder):
+    try:
+        # restore old
+        deploy_cache = read_deploy(current_folder + "\\deploy.cfg")
+        if deploy_cache == "":
+            return True
+        
+        target_file_1 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_1.wav"
+        target_file_2 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_2.wav"
+        target_file_3 = game_path + "Sounds\\Music_Theme_" + deploy_cache + "_3.wav"
+        remove_with_restore(target_file_1)
+        remove_with_restore(target_file_2)
+        remove_with_restore(target_file_3)
+    except:
+        return False
     return True
+
+# ========================================== assistant function
 
 def copy_with_backups(target, origin):
     if os.path.exists(target):
         if not os.path.exists(target + ".bak"):
             os.rename(target, target+ ".bak")
+        else:
+            os.remove(target)
     shutil.copyfile(origin, target)
 
 def remove_with_restore(target):
-    os.remove(target)
+    if os.path.exists(target):
+        os.remove(target)
     if os.path.exists(target+".bak"):
         os.rename(target+".bak", target)
+
+def record_deploy(file, value):
+    with opne(file, "w", encoding="utf-8") as f:
+        f.write(value)
+
+def read_deploy(file):
+    if not os.path.exists(file):
+        return ""
+    
+    with opne(file, "r", encoding="utf-8") as f:
+        return f.read()
